+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 [Full diff](https://github.com/sider/runners/compare/0.40.5...HEAD)
 
 - **ESLint** Add `eslint-plugin-cypress` to pre-installed packages [#1936](https://github.com/sider/runners/pull/1936)
+- **RuboCop** Revert rubydoc.info links [#1942](https://github.com/sider/runners/pull/1942)
 
 ## 0.40.5